refactor: using `Call` macro in `Balances`
https://build.w3d.community/courses/Rust_State_Machine/Section_6/Lesson_2_Adding_Call_Macro_to_Balances.md?lang=en

1. If you haven't, move your transfer function into its own impl<T: Config> Pallet<T>.
We only want to apply the macro to this one function, so we need to isolate it from the other functions which are not
meant to be callable.
2. Add the #[macros::call] attribute over this new impl<T: Config> Pallet<T>.
3. Delete your existing enum Call.
4. Delete your existing implementation of Dispatch for Pallet.
5. Then, in your main.rs file, change instances of balances::Call::Transfer to balances::Call::transfer with a lowercase t.

MORE:
Rust provides the command cargo expand which allows you to output the generated rust code after all macros have been applied to your project.

cargo expand > out.rs